+++ cli/tests/test_git_push.rs
@@ -50,7 +50,7 @@ fn set_up() -> (TestEnvironment, PathBuf) {
 fn test_git_push_nothing() {
     let (test_env, workspace_root) = set_up();
     // Show the setup. `insta` has trouble if this is done inside `set_up()`
-    let stdout = test_env.jj_cmd_success(&workspace_root, &["branch", "list", "--all"]);
+    let stdout = test_env.jj_cmd_success(&workspace_root, &["branch", "list", "--all-remotes"]);
     insta::assert_snapshot!(stdout, @r###"
     branch1: lzmmnrxq 45a3aa29 (empty) description 1
       @origin: lzmmnrxq 45a3aa29 (empty) description 1
@@ -80,7 +80,7 @@ fn test_git_push_current_branch() {
     test_env.jj_cmd_ok(&workspace_root, &["branch", "create", "my-branch"]);
     test_env.jj_cmd_ok(&workspace_root, &["describe", "-m", "foo"]);
     // Check the setup
-    let stdout = test_env.jj_cmd_success(&workspace_root, &["branch", "list", "--all"]);
+    let stdout = test_env.jj_cmd_success(&workspace_root, &["branch", "list", "--all-remotes"]);
     insta::assert_snapshot!(stdout, @r###"
     branch1: lzmmnrxq 19e00bf6 (empty) modified branch1 commit
       @origin (ahead by 1 commits, behind by 1 commits): lzmmnrxq hidden 45a3aa29 (empty) description 1
@@ -104,7 +104,7 @@ fn test_git_push_current_branch() {
       Move branch branch2 from 8476341eb395 to 10ee3363b259
       Add branch my-branch to 10ee3363b259
     "###);
-    let stdout = test_env.jj_cmd_success(&workspace_root, &["branch", "list", "--all"]);
+    let stdout = test_env.jj_cmd_success(&workspace_root, &["branch", "list", "--all-remotes"]);
     insta::assert_snapshot!(stdout, @r###"
     branch1: lzmmnrxq 19e00bf6 (empty) modified branch1 commit
       @origin (ahead by 1 commits, behind by 1 commits): lzmmnrxq hidden 45a3aa29 (empty) description 1
@@ -256,7 +256,7 @@ fn test_git_push_locally_created_and_rewritten() {
     // Rewrite it and push again, which would fail if the pushed branch weren't
     // set to "tracking"
     test_env.jj_cmd_ok(&workspace_root, &["describe", "-mlocal 2"]);
-    let stdout = test_env.jj_cmd_success(&workspace_root, &["branch", "list", "--all"]);
+    let stdout = test_env.jj_cmd_success(&workspace_root, &["branch", "list", "--all-remotes"]);
     insta::assert_snapshot!(stdout, @r###"
     branch1: lzmmnrxq 45a3aa29 (empty) description 1
       @origin: lzmmnrxq 45a3aa29 (empty) description 1
@@ -283,7 +283,7 @@ fn test_git_push_multiple() {
     test_env.jj_cmd_ok(&workspace_root, &["branch", "create", "my-branch"]);
     test_env.jj_cmd_ok(&workspace_root, &["describe", "-m", "foo"]);
     // Check the setup
-    let stdout = test_env.jj_cmd_success(&workspace_root, &["branch", "list", "--all"]);
+    let stdout = test_env.jj_cmd_success(&workspace_root, &["branch", "list", "--all-remotes"]);
     insta::assert_snapshot!(stdout, @r###"
     branch1 (deleted)
       @origin: lzmmnrxq 45a3aa29 (empty) description 1
@@ -369,7 +369,7 @@ fn test_git_push_multiple() {
       Force branch branch2 from 8476341eb395 to 15dcdaa4f12f
       Add branch my-branch to 15dcdaa4f12f
     "###);
-    let stdout = test_env.jj_cmd_success(&workspace_root, &["branch", "list", "--all"]);
+    let stdout = test_env.jj_cmd_success(&workspace_root, &["branch", "list", "--all-remotes"]);
     insta::assert_snapshot!(stdout, @r###"
     branch2: yqosqzyt 15dcdaa4 (empty) foo
       @origin: yqosqzyt 15dcdaa4 (empty) foo
@@ -757,7 +757,7 @@ fn test_git_push_conflicting_branches() {
     test_env.jj_cmd_ok(&workspace_root, &["branch", "create", "branch2"]);
     test_env.jj_cmd_ok(&workspace_root, &["git", "fetch"]);
     insta::assert_snapshot!(
-        test_env.jj_cmd_success(&workspace_root, &["branch", "list", "--all"]), @r###"
+        test_env.jj_cmd_success(&workspace_root, &["branch", "list", "--all-remotes"]), @r###"
     branch1: lzmmnrxq 45a3aa29 (empty) description 1
       @origin: lzmmnrxq 45a3aa29 (empty) description 1
     branch2 (conflicted):
@@ -837,7 +837,7 @@ fn test_git_push_tracked_vs_all() {
     test_env.jj_cmd_ok(&workspace_root, &["branch", "delete", "branch2"]);
     test_env.jj_cmd_ok(&workspace_root, &["branch", "untrack", "branch1@origin"]);
     test_env.jj_cmd_ok(&workspace_root, &["branch", "create", "branch3"]);
-    let stdout = test_env.jj_cmd_success(&workspace_root, &["branch", "list", "--all"]);
+    let stdout = test_env.jj_cmd_success(&workspace_root, &["branch", "list", "--all-remotes"]);
     insta::assert_snapshot!(stdout, @r###"
     branch1: vruxwmqv a25f24af (empty) moved branch1
     branch1@origin: lzmmnrxq 45a3aa29 (empty) description 1
@@ -859,7 +859,7 @@ fn test_git_push_tracked_vs_all() {
 
     // Untrack the last remaining tracked branch.
     test_env.jj_cmd_ok(&workspace_root, &["branch", "untrack", "branch2@origin"]);
-    let stdout = test_env.jj_cmd_success(&workspace_root, &["branch", "list", "--all"]);
+    let stdout = test_env.jj_cmd_success(&workspace_root, &["branch", "list", "--all-remotes"]);
     insta::assert_snapshot!(stdout, @r###"
     branch1: vruxwmqv a25f24af (empty) moved branch1
     branch1@origin: lzmmnrxq 45a3aa29 (empty) description 1